==> BuildingDB/02_CourtListener.ipynb <==
import polars as pl
import requests
import os
import re

API_KEY = os.environ["CL_API_KEY"]
# defining a params dict for the parameters to be sent to the API
header =  {"Authorization": f"Token {API_KEY}"}

df = pl.read_csv("./Cases/WestLawMatch.csv")

def get_cluster_id(absolute_url):
    pattern = r'/opinion/(\d+)/'

    match = re.search(pattern, absolute_url)
    if match:
        return match.group(1)

df = df.with_columns(
    pl.col("OpinionURL").map_elements(get_cluster_id).alias("ClusterID")
)

df.head()

def get_opinions_from_cluster(cluster_id, header = header):

    URL = f"https://www.courtlistener.com/api/rest/v4/clusters/{cluster_id}"

        # sending get request and saving the response as response object
    r = requests.get(url = URL, headers = header)

    if r.status_code == 200:

        response = r.json()
        return response
    
    else:

        return { None }

get_opinions_from_cluster("7331944")

df = df.with_columns(
    pl.col("ClusterID").map_elements(get_opinions_from_cluster).alias("opinions")
)

cl = df.select(["OpinionURL", "CourtListenerCaseName", "Citation","ClusterID", "opinions"]).with_columns(
    pl.col("opinions").map_elements(lambda x: x["docket_id"]).alias("DocketID"),
    pl.col("opinions").map_elements(lambda x: x["sub_opinions"]).alias("SubOpinions"),
    pl.col("opinions").map_elements(lambda x: x["precedential_status"]).alias("PrecedentialStatus")
).drop("opinions")

cl = cl.explode("SubOpinions")

cl.head()

def get_actual_opinion(opinion_id):

    URL = opinion_id

    # sending get request and saving the response as response object
    r = requests.get(url = URL, headers = header)

    if r.status_code == 200:

        response = r.json()
            
        return response
        
    else:

        return None

cl = cl.with_columns(
    pl.col("SubOpinions").map_elements(get_actual_opinion).alias("Opinion_API")
)

def get_raw_opinion(opinion_data: dict):

    # Priority order for locating the opinion text
    document_priority = [
        "html_with_citations",
        "html_columbia",
        "html_lawbox",
        "xml_harvard",
        "html_anon_2020",
        "html",
        "plain_text",
    ]

    # Find the first non-empty document in the priority list
    document = None
    for key in document_priority:
        # We check if the key is in the dict and not None or empty
        if key in opinion_data and opinion_data[key]:
            document = opinion_data[key]
            break

    return document

cl = cl.with_columns(
    pl.col("Opinion_API").map_elements(lambda x: x["type"]).alias("OpinionType"),
    pl.col("Opinion_API").map_elements(get_raw_opinion).alias("Document")
).drop("Opinion_API")

cl.write_csv("./Cases/CourtListenerOpinions.csv")

def get_courts_from_docket(docket_id, header = header):

    docket_id = str(docket_id)

    URL = f"https://www.courtlistener.com/api/rest/v4/dockets/{docket_id}"

    # sending get request and saving the response as response object
    r = requests.get(url = URL, headers = header)

    if r.status_code == 200:

        response = r.json()
        return response
    
    else:

        return { None }

def get_courts_from_court_id(court_url, header = header):

    URL = court_url

    # sending get request and saving the response as response object
    r = requests.get(url = URL, headers = header)

    if r.status_code == 200:

        response = r.json()
        return response
    
    else:

        return { None }

df = pl.read_csv("./Cases/CourtListenerOpinions.csv")

df = df.with_columns(
    pl.col("DocketID").map_elements(get_courts_from_docket).alias("Docket_API")
)

df = df.with_columns(
    pl.col("Docket_API").map_elements(lambda x: x["court_id"]).alias("CourtID"),
    pl.col("Docket_API").map_elements(lambda x: x["court"]).alias("CourtURL")
)

df = df.with_columns(
    pl.col("CourtURL").map_elements(get_courts_from_court_id).alias("Court_API")
)

df.with_columns(
    pl.col("Court_API").map_elements(lambda x: x["full_name"]).alias("Court")
).drop(["Docket_API", "Court_API"]).write_csv("./Cases/CourtListenerWithCourt.csv")
