==> BuildingDB/02a_WestLawNoMatch.ipynb <==
import polars as pl
import os
from util import pdf_to_text, best_fuzzy_match

ms_cases = [c.replace(".pdf", "") for c in os.listdir("Cases/ManuallyScrapedCases/")]
df = pl.read_csv("./Cases/WestLawNotMatch.csv")
df.head()

df = df.with_columns(
    pl.col("Title").map_elements(lambda x: best_fuzzy_match(x, choices=ms_cases)[0]+".pdf").alias("Matched"),
    pl.col("Title").map_elements(lambda x: best_fuzzy_match(x, choices=ms_cases)[1]).alias("MatchedScore")
).sort("MatchedScore")

df = df.filter( pl.col("MatchedScore") > 49).with_columns(
    pl.col("Matched").map_elements(lambda x: pdf_to_text("./Cases/ManuallyScrapedCases/" + x)).alias("Opinion")
)

df.write_csv("./Cases/WLNM_Opinion.csv")
