==> BuildingDB/07_CheckingForError.ipynb <==
import polars as pl
from neo4j import GraphDatabase

URI = "bolt://localhost:7687"
AUTH = ("neo4j", "fairusecases")

def get_all_opinions(tx):
    
    results = tx.run("""
        MATCH (c:Case)-[]-(o:Opinion)
        RETURN c.WestLawCaseName as Title, o.Document as Document
        """
    )
         
    return results.to_df()


with GraphDatabase.driver(URI, auth=AUTH) as driver:
    with driver.session(database="neo4j") as session:
            df = session.execute_read(get_all_opinions)

df = pl.from_pandas(df)

df.with_columns(
    pl.col("Document").map_elements(len).alias("DocLength")
).sort(by = "DocLength")

df.group_by("Title").count().filter(pl.col("count") != 1)["Title"].to_list()
