==> BuildingDB/08_CreatingSummary.ipynb <==
from langchain_google_genai import ChatGoogleGenerativeAI
from langchain_core.prompts import ChatPromptTemplate
import os
from neo4j import GraphDatabase
from dotenv import load_dotenv
import polars as pl

dotenv_path = os.path.abspath("../.env")  # Adjust the path accordingly
load_dotenv(dotenv_path)

URI = "bolt://localhost:7687"
AUTH = ("neo4j", "fairusecases")

os.environ["GOOGLE_API_KEY"] = os.environ["GEMINI_API"]

# Initialize the LLM (ensure your API key is set in your environment)
llm = ChatGoogleGenerativeAI(
    model="gemini-2.0-flash",
    max_tokens=None,
    timeout=None
)

fair_use_summary_prompt = ChatPromptTemplate.from_messages(
    [
        (
            "system",
            """You are a legal analyst. Read the following legal opinion and summarize its significance with respect to its Fair Use ruling. Focus on the key points, include a general overview of the facts of the case and the context, as well as the implications regarding Fair Use.\n\n""",
        ),
        ("human", "Here is the legal opinion: {legal_opinion}"),
    ]
)

fair_use_summary_chain = fair_use_summary_prompt | llm

def get_opinions(tx):
    
    results = tx.run("""
        MATCH (o:Opinion)
        WHERE o.Summary IS NULL
        RETURN o.Document as Document
        """
    )
         
    return pl.from_pandas(results.to_df())


with GraphDatabase.driver(URI, auth=AUTH) as driver:
    with driver.session(database="neo4j") as session:
            df = session.execute_read(get_opinions)

df.describe

def create_opinion_summary(tx, document, summary):
    
    tx.run("""
        MATCH (o:Opinion {Document: $document})
        SET o.Summary = $summary
        """, document = document, summary = summary
    )
         

with GraphDatabase.driver(URI, auth=AUTH) as driver:
    with driver.session(database="neo4j") as session:
            for d in df.iter_rows(named = True):
                summary = fair_use_summary_chain.invoke({"legal_opinion": d["Document"]}).content
                session.execute_write(create_opinion_summary, d["Document"], summary)
